refactor. 프론트 페이지 폴더구조 변경
- 페이지 컴포넌트 옆에 style 파일 설정
- scss 활용
