+++ frontend/src/index.js
@@ -1,4 +1,4 @@
-import './style/style.css';
+import './style/style.scss';
 import App from './App.js';
 
 const $root = document.querySelector('#root');
