+++ frontend/src/store/page.js
@@ -1,4 +1,7 @@
-import HomePage from '../pages/HomePage.js';
+import HomePage from '../pages/HomePage';
 import { initState } from '../utils/globalObserver.js';
 
-export const pageState = initState({ key: 'page', defaultValue: { Page: HomePage, direction: '' } });
+export const pageState = initState({
+  key: 'page',
+  defaultValue: { Page: HomePage, direction: '' },
+});
